# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TimeDive API — history learning through personalized AI stories
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  displayName: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  token: string;
  /** @minLength 8 */
  password: string;
}

/**
 * @nullable
 */
export type AuthUserAgeMode = typeof AuthUserAgeMode[keyof typeof AuthUserAgeMode] | null;


export const AuthUserAgeMode = {
  kid: 'kid',
  teen: 'teen',
  adult: 'adult',
} as const;

export interface AuthUser {
  id: number;
  email: string;
  displayName: string;
  onboardingComplete: boolean;
  /** @nullable */
  ageMode?: AuthUserAgeMode;
  /** @nullable */
  gradeLevel?: number | null;
}

/**
 * @nullable
 */
export type UserProfileAgeMode = typeof UserProfileAgeMode[keyof typeof UserProfileAgeMode] | null;


export const UserProfileAgeMode = {
  kid: 'kid',
  teen: 'teen',
  adult: 'adult',
} as const;

export interface UserProfile {
  id: number;
  email: string;
  displayName: string;
  /** @nullable */
  ageMode?: UserProfileAgeMode;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gradeLevel?: number | null;
  onboardingComplete: boolean;
  recapEmailOptIn?: boolean;
  createdAt: string;
  /** @nullable */
  lastLoginAt?: string | null;
}

export type ProfileUpdateAgeMode = typeof ProfileUpdateAgeMode[keyof typeof ProfileUpdateAgeMode];


export const ProfileUpdateAgeMode = {
  kid: 'kid',
  teen: 'teen',
  adult: 'adult',
} as const;

export interface ProfileUpdate {
  displayName?: string;
  ageMode?: ProfileUpdateAgeMode;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gradeLevel?: number | null;
  onboardingComplete?: boolean;
  recapEmailOptIn?: boolean;
}

export interface UserPreferences {
  userId: number;
  sports?: string[];
  videoGames?: string[];
  movieGenres?: string[];
  hobbies?: string[];
  books?: string[];
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
}

export interface PreferencesUpdate {
  sports?: string[];
  videoGames?: string[];
  movieGenres?: string[];
  hobbies?: string[];
  books?: string[];
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
}

export interface Topic {
  id: number;
  eraName: string;
  category: string;
  /** 1=shallow/recent, 10=deep/ancient */
  depthLevel: number;
  description: string;
  /** @nullable */
  gradeRangeMin?: number | null;
  /** @nullable */
  gradeRangeMax?: number | null;
  /** Whether the current user has completed this topic */
  isCompleted?: boolean;
}

export interface TopicDetail {
  id: number;
  eraName: string;
  category: string;
  depthLevel: number;
  description: string;
  coreFacts: string[];
  /** @nullable */
  gradeRangeMin?: number | null;
  /** @nullable */
  gradeRangeMax?: number | null;
  isCompleted?: boolean;
  hasStory?: boolean;
}

export interface Story {
  id: number;
  topicId: number;
  storyText: string;
  funFacts: string;
  generatedAt: string;
}

export interface GenerateStoryInput {
  forceRegenerate?: boolean;
}

export interface CategoryProgress {
  category: string;
  total: number;
  completed: number;
}

export interface RecentCompletion {
  topicId: number;
  eraName: string;
  completedAt: string;
}

export interface ProgressSummary {
  totalTopics: number;
  completedTopics: number;
  completedByCategory: CategoryProgress[];
  recentCompletions: RecentCompletion[];
}

export interface CompleteTopicInput {
  topicId: number;
}

export interface Badge {
  id: number;
  name: string;
  description: string;
  category: string;
  icon: string;
  earnedAt: string;
}

export interface CompletionResult {
  newBadges: Badge[];
  totalCompleted: number;
}

export interface DashboardSummary {
  totalTopics: number;
  completedTopics: number;
  recentTopics: RecentCompletion[];
  suggestedNext: Topic[];
  badgeCount?: number;
  /** @nullable */
  lastLoginAt?: string | null;
}