# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TimeDive API — history learning through personalized AI stories
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get currently authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent (always 200 to prevent enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /topics/{id}:
    get:
      operationId: getTopic
      tags: [topics]
      summary: Get a single topic with core facts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Topic detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stories/{topicId}:
    get:
      operationId: getStory
      tags: [stories]
      summary: Get cached story for the current user and topic
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
        "404":
          description: Story not yet generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stories/{topicId}/generate:
    post:
      operationId: generateStory
      tags: [stories]
      summary: Generate (or regenerate) a personalized story for a topic
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStoryInput"
      responses:
        "200":
          description: Generated story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
        "404":
          description: Topic not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth schemas
    RegisterInput:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        displayName:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    AuthUser:
      type: object
      required: [id, email, displayName, onboardingComplete]
      properties:
        id:
          type: number
        email:
          type: string
        displayName:
          type: string
        onboardingComplete:
          type: boolean
        ageMode:
          type: ["string", "null"]
          enum: [kid, teen, adult, null]
        gradeLevel:
          type: ["integer", "null"]

    # User profile schemas
...
    TopicDetail:
      type: object
      required: [id, eraName, category, depthLevel, description, coreFacts]
      properties:
        id:
          type: number
        eraName:
          type: string
        category:
          type: string
        depthLevel:
          type: number
        description:
          type: string
        coreFacts:
          type: array
          items:
            type: string
        gradeRangeMin:
          type: ["integer", "null"]
        gradeRangeMax:
          type: ["integer", "null"]
        isCompleted:
          type: boolean
        hasStory:
          type: boolean

    # Story schemas
    Story:
      type: object
      required: [id, topicId, storyText, funFacts, generatedAt]
      properties:
        id:
          type: number
        topicId:
          type: number
        storyText:
          type: string
        funFacts:
          type: string
        generatedAt:
          type: string
          format: date-time
    GenerateStoryInput:
      type: object
      properties:
        forceRegenerate:
          type: boolean

    # Progress schemas
...