# lib/api-zod/src/generated/types/generateStoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TimeDive API — history learning through personalized AI stories
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateStoryInput {
  forceRegenerate?: boolean;
}