# lib/api-zod/src/generated/types/topicDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TimeDive API — history learning through personalized AI stories
 * OpenAPI spec version: 0.1.0
 */

export interface TopicDetail {
  id: number;
  eraName: string;
  category: string;
  depthLevel: number;
  description: string;
  coreFacts: string[];
  /** @nullable */
  gradeRangeMin?: number | null;
  /** @nullable */
  gradeRangeMax?: number | null;
  isCompleted?: boolean;
  hasStory?: boolean;
}